// sensors/src/app/models/sensor.ts
export interface Sensor {
    ConnectionType: any;
    InstallDate: string;
    ComponentOk: number;
    DeviceTypeHebrew: string;
    DeviceId: string;
    DeviceType: string;
    WebSiteDeviceName: string;
    LastReportDate: string;
    Picture: string;
    ManufacturerName: string;
    ReceptionLevel: string;
    MAC: string;
}